Adding documentation
Let's add the documentation and update the categories file
