+++ documentation/README.md
@@ -1,0 +1,39 @@
+# AI Construct PDF Splitter Documentation
+
+Welcome to the documentation for the AI Construct PDF Splitter tool. This documentation will help you understand, use, and extend the application.
+
+## Available Documentation
+
+| Document | Description |
+|----------|-------------|
+| [Main Documentation](documentation.md) | Comprehensive guide covering all aspects of the application |
+| [Quick Start Guide](quick_start_guide.md) | Get up and running quickly with the essential steps |
+| [Developer Guide](developer_guide.md) | Technical information for developers who want to extend or modify the application |
+| [Category File Guide](category_file_guide.md) | Detailed instructions for creating and customizing category files |
+
+## About AI Construct PDF Splitter
+
+The AI Construct PDF Splitter is a specialized tool designed to process construction specification documents (lastenboeken), particularly targeting non-VMSW (Flemish Social Housing Company) documents. The application analyzes PDF documents, extracts their table of contents, categorizes the content using Google's Gemini AI, and splits the document into separate PDFs based on identified categories.
+
+## Key Features
+
+- Automatic extraction of document structure from PDF files
+- AI-powered categorization of document sections
+- Creation of separate PDF files for each construction category
+- User-friendly graphical interface
+- Command-line interface for automation
+- Support for custom category definitions
+
+## Getting Started
+
+For first-time users, we recommend starting with the [Quick Start Guide](quick_start_guide.md), which will walk you through the basic installation and usage steps.
+
+For a complete understanding of all features, consult the [Main Documentation](documentation.md).
+
+## For Developers
+
+If you're interested in extending or modifying the application, refer to the [Developer Guide](developer_guide.md) for technical details about the code structure and architecture.
+
+## Creating Custom Categories
+
+To create your own custom category definitions, see the [Category File Guide](category_file_guide.md) for detailed instructions and examples.
